% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.r
\name{args_or_condo}
\alias{args_or_condo}
\title{Constructs a string of arguments for ssh exection on Rhea}
\usage{
args_or_condo(port = 55555, verbose = FALSE)
}
\arguments{
\item{port}{The port for server connection}

\item{verbose}{Should the args be printed?}
}
\description{
Constructs a string of arguments for ssh exection on Rhea
}
