% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.r
\name{args_rhea}
\alias{args_rhea}
\title{Constructs a string of arguments for ssh exection on Rhea}
\usage{
args_rhea(port = 55555)
}
\arguments{
\item{port}{The port for server connection}
}
\description{
Constructs a string of arguments for ssh exection on Rhea
}
