% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.r
\name{here_doc}
\alias{here_doc}
\title{Wraps a script with shell commands to write it to a file.
Sometimes referred to as a "here document"}
\usage{
here_doc(script, file_name, eof)
}
\arguments{
\item{script}{A vector of strings, usually each a shell command.}

\item{file_name}{A string giving the file name to use. Careful not to
conflict with an existing file name as it will be silently deleted and
overwritten.}

\item{eof}{A string to use for end-of-file. Must not occurr within the
script and must be unique among potential nested uses of this function
for a script.}
}
\value{
A vector of strings, each string a line of the script.
}
\description{
Wraps a script with shell commands to write it to a file.
Sometimes referred to as a "here document"
}
