% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launches a persistent pbdR server on a cluster}
\usage{
launch(nodes, npernode = 16, server = "rhea.ccs.ornl.gov",
  modules = c("r"), user = Sys.getenv("USER"), account = NULL,
  walltime = "01:00:00", rwd = "~/")
}
\arguments{
\item{nodes}{The number of nodes to allocate for the server.}

\item{npernode}{The number of MPI ranks (R sessions) to start per node.}

\item{server}{Hostname to run the pbdR server, resolvable on your
platform.}

\item{modules}{A vector of strings giving the modules to load before
starting the server.}

\item{user}{Username for login.}

\item{account}{Account used for your cluster allocation}

\item{walltime}{A string "hh:mm:ss" giving the maximum length of time
in hours:minutes:seconds before server timeout shutdown.}

\item{rwd}{Remote working directory as a string.}
}
\value{
A shell script as a vector of strings.
}
\description{
Launches a persistent pbdR server on a cluster
}
