% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchr-package.r
\docType{package}
\name{launchr-package}
\alias{launchr-package}
\title{launchr}
\description{
TODO
}
\references{
Project URL: \url{https://github.com/RBigData/launchr}
}
\author{
George Ostrouchov
}
\keyword{package}
