% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.r
\name{lnode_script}
\alias{lnode_script}
\title{Constructs a shell script as a vector of strings. Its purpose is to run
in a login node shell and submit a pbdR server batch job that
runs for a specified time or until closed. It also opens an ssh tunnel
from the login node to the running server head node.}
\usage{
lnode_script(fn, port, rwd = "~/", ...)
}
\arguments{
\item{fn}{A list of file names that will be created on the
server. Normally the defaults specified in \code{launch()}
function are fine.}

\item{port}{The port used by the server.}

\item{rwd}{Remote working directory as a string.}

\item{...}{Unmatched arguments are passed down to \code{pbs_default()}}
}
\value{
A shell script as a vector of strings.
}
\description{
Modification for platform-specific parameters should be done with a
function via the FUN parameter of the \code{launch()} function
rather than rewriting \code{lnode_script()}.
}
