% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.r
\name{preload_rhea}
\alias{preload_rhea}
\title{Constructs a shell script as a vector of strings. Its purpose is to run
on a login node of rhea.ccs.ornl.gov, submit a pbdR server batch job that
runs for a specified time or until closed. It also opens an ssh tunnel
from the login node to the running server head node.}
\usage{
preload_rhea(nodes = 1, npernode = 16, walltime = "01:00:00",
  user = "ost", machine = "rhea.ccs.ornl.gov", port = 55555,
  account = "gen011", modules = c("r"), rwd = "~")
}
\arguments{
\item{nodes}{The number of nodes to allocate for the server.}

\item{npernode}{The number of MPI ranks (R sessions) to start per node.}

\item{walltime}{A string "hh:mm:ss" giving the maximum length of time
in hours:minutes:seconds before server timeout shutdown.}

\item{user}{Username for login.}

\item{machine}{Hostname to run the server, resolvable on your platform.}

\item{port}{The port used by the server.}

\item{account}{Account used for your cluster allocation}

\item{modules}{A vector of strings giving the modules to load before
starting the server.}

\item{rwd}{Remote working directory as a string.}
}
\value{
A shell script as a vector of strings.
}
\description{
Constructs a shell script as a vector of strings. Its purpose is to run
on a login node of rhea.ccs.ornl.gov, submit a pbdR server batch job that
runs for a specified time or until closed. It also opens an ssh tunnel
from the login node to the running server head node.
}
