% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.r
\name{ssh_args}
\alias{ssh_args}
\title{Constructs a string of arguments for local ssh to forward a port}
\usage{
ssh_args(port)
}
\arguments{
\item{port}{The port for server connection}
}
\value{
A string.
}
\description{
Constructs a string of arguments for local ssh to forward a port
}
